Use inttypes.h macros in libpcp to address Win32 build warnings.
Moving toward a warning free build on Windows, the main remaining issue
is the following class of problem:

events.c: In function '__pmDumpEventRecords':
events.c:120: warning: unknown conversion type character 'l' in format
events.c:120: warning: too many arguments for format
events.c:124: warning: unknown conversion type character 'l' in format
events.c:124: warning: too many arguments for format

googling suggests this is because the Windows printf implementation is
not conforming to the same standards as most others players (%llu and
%lld are the typical sources of the above warning).  Evidentally, all
platforms do have inttypes.h which seems to be recommended for solving
this, Wikipedia states "ISO C99 includes the inttypes.h header file that
includes a number of macros for use in platform-independent printf coding"
in http://en.wikipedia.org/wiki/Printf#Format_placeholders ... this does
indeed address the problem.

+++ src/libpcp/src/events.c
@@ -15,6 +15,7 @@
  * License for more details.
  */
 
+#include <inttypes.h>
 #include "pmapi.h"
 #include "impl.h"
 
@@ -117,11 +118,11 @@ __pmDumpEventRecords(FILE *f, pmValueSet *vsp)
 		    break;
 		case PM_TYPE_64:
 		    memcpy((void *)&atom.ll, (void *)vbuf, sizeof(atom.ll));
-		    fprintf(f, " = %lli", (long long)atom.ll);
+		    fprintf(f, " = %"PRIi64, atom.ll);
 		    break;
 		case PM_TYPE_U64:
 		    memcpy((void *)&atom.ull, (void *)vbuf, sizeof(atom.ull));
-		    fprintf(f, " = %llu", (unsigned long long)atom.ull);
+		    fprintf(f, " = %"PRIu64, atom.ull);
 		    break;
 		case PM_TYPE_FLOAT:
 		    memcpy((void *)&atom.f, (void *)vbuf, sizeof(atom.f));
